pyproject.toml
[project]
name = "instapybot"
version = "0.1.0"
description = "Insta Py Bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.66.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
requires = ["pyright"]
